# M6-Challenge-Garcia-Zulymar/openapi.yaml
#20230322182847
#http://localhost:8080/v3/api-docs

    {
      "openapi": "3.0.1",
      "info": {
        "title": "OpenAPI definition",
        "version": "v0"
      },
      "servers": [
        {
          "url": "http://localhost:8080",
          "description": "Generated server url"
        }
      ],
      "paths": {
        "/customers": {
          "get": {
            "tags": [
                "customer-controller"
            ],
            "operationId": "getAllCustomers",
            "responses": {
              "200": {
                "description": "OK",
                "content": {
                  "*/*": {
                    "schema": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Customer"
                      }
                    }
                  }
                }
              }
            }
          },
          "put": {
            "tags": [
                "customer-controller"
            ],
            "operationId": "updateCustomer",
            "requestBody": {
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Customer"
                  }
                }
              },
              "required": true
            },
            "responses": {
              "204": {
                "description": "No Content"
              }
            }
          },
          "post": {
            "tags": [
                "customer-controller"
            ],
            "operationId": "addCustomer",
            "requestBody": {
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Customer"
                  }
                }
              },
              "required": true
            },
            "responses": {
              "201": {
                "description": "Created",
                "content": {
                  "*/*": {
                    "schema": {
                      "$ref": "#/components/schemas/Customer"
                    }
                  }
                }
              }
            }
          }
        },
        "/customers/{id}": {
          "get": {
            "tags": [
                "customer-controller"
            ],
            "operationId": "getCustomerByID",
            "parameters": [
              {
                "name": "id",
                "in": "path",
                "required": true,
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            ],
            "responses": {
              "200": {
                "description": "OK",
                "content": {
                  "*/*": {
                    "schema": {
                      "$ref": "#/components/schemas/Customer"
                    }
                  }
                }
              }
            }
          },
          "delete": {
            "tags": [
                "customer-controller"
            ],
            "operationId": "deleteCustomer",
            "parameters": [
              {
                "name": "id",
                "in": "path",
                "required": true,
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            ],
            "responses": {
              "204": {
                "description": "No Content"
              }
            }
          }
        },
        "/customers/state/{state}": {
          "get": {
            "tags": [
                "customer-controller"
            ],
            "operationId": "getCustomersByState",
            "parameters": [
              {
                "name": "state",
                "in": "path",
                "required": true,
                "schema": {
                  "type": "string"
                }
              }
            ],
            "responses": {
              "200": {
                "description": "OK",
                "content": {
                  "*/*": {
                    "schema": {
                      "uniqueItems": true,
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Customer"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "components": {
        "schemas": {
          "Customer": {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer",
                "format": "int32"
              },
              "firstName": {
                "type": "string"
              },
              "lastName": {
                "type": "string"
              },
              "email": {
                "type": "string"
              },
              "company": {
                "type": "string"
              },
              "phone": {
                "type": "string"
              },
              "address_1": {
                "type": "string"
              },
              "address_2": {
                "type": "string"
              },
              "city": {
                "type": "string"
              },
              "state": {
                "type": "string"
              },
              "postalCode": {
                "type": "string"
              },
              "country": {
                "type": "string"
              }
            }
          }
        }
      }
    }